Update index.adoc
I'm not sure if this commit actually makes the doc less accurate or more. Depends on whether or not an ITS something that takes the form of a YAML file, or something that is defined by one. After looking at all our docs, I thought it could go either way. But I think it might help users if we define the ITS as the YAML file, since that makes it more tangible.