Rework supply chain metadata documentation
In order to be able to more easily highlight convenient commands and
usage for the generated SLSA provenance, I wanted to restructure the
documentation. It now provides an overview of the supply chain metadata,
how to see the associated metadata. Within each section we can now be
more focused on how to leverage the metadata to answer questions and
enable auditability.

I consolidated the process for identifying the container images and put
this outside of the metadata directory as it is more generically
valuable.

Signed-off-by: arewm <arewm@users.noreply.github.com>