docs(RELEASE-1320): update tenant pipeline example pipelineRef
The pipelineRef in the tenant pipeline never really made much sense. It
suggested using a managed pipeline, which would not work (and many users
made this mistake). This updates the example to something that should be
more obviously an example and not a real pipeline.

We also moved the pipeline definitions in release-service-catalog, so
the old reference won't even exist going forward. That is another reason
to change it.

If a working tenant pipeline is created, we can update the docs to use
that as the example.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>