KONFLUX-6912: Generate API Reference Docs
Initial attempt to generate reference documentation for Konflux APIs.
This adds scripting to generate reference docs for Kubernetes CRDs
that are added as part of a Konflux instance. The process utilizes the
`crd-ref-docs` tool to generate asciidoc directly from source code.

Generation of equivalent reference docs for upstream dependencies, such
as Tekton, is out of scope for this iteration. Future improvements may
consider providing links to upstream sources.

Signed-off-by: Adam Kaplan <adam.kaplan@redhat.com>