KONFLUX-6912: Generate API Reference Docs
Initial attempt to generate reference documentation for Konflux APIs.
This adds scripting to generate reference docs for Kubernetes CRDs
that are added as part of a Konflux instance. The process utilizes the
`crd-ref-docs` tool to generate asciidoc directly from source code.
The format of the generated asciidoc can be adjusted by updating the
files in the `api-gen/templates` directory, which are Go-based text
templates. Refer to the crd-ref-docs codebase for available template
functions [1].

Generation of equivalent reference docs for upstream dependencies, such
as Tekton, is out of scope for this iteration. Future improvements may
consider providing links to upstream sources. Annotations and labels that
alter the behavior of upstream APIs are likewise out of scope for this
iteration, but can be considered as a future/follow-up improvement.

Note - some Enterprise Contract validations contains regex that are not
correctly escaped for asciidoc. This is a known issue with crd-ref-docs
and requires manual intervention to make `npm run build` happy [2].

[1] https://github.com/elastic/crd-ref-docs
[2] https://github.com/elastic/crd-ref-docs/issues/137

Signed-off-by: Adam Kaplan <adam.kaplan@redhat.com>