Unify the Getting Started pages
The information contained in the "Getting started with the CLI" page is
no longer relevant, so it is being dropped.

The current "Getting started with Konflux" is being moved to a top
level page, and a partial related to CLI configuration is being
introduced to enable easier updates in the future.

Signed-off-by: Bruno Pimentel <bpimente@redhat.com>