Move getting-started content to the ROOT module
This is part of a major refactoring that aims to move all content nested
under root to a specific module, according to the best practices
established by Antora (see their official docs repo¹)

The page beind moved is a top-level page, so it should be placed directly
under the ROOT module. The nav.adoc was also placed in the right place.

¹https://gitlab.com/antora/antora/-/tree/main/docs/modules

Signed-off-by: Bruno Pimentel <bpimente@redhat.com>