prefetching: Remove mentions of Cachi2
1.
Replace all mentions of Cachi2 with Hermeto in the documentation to reflect
the project name change. This ensures consistency and accuracy in the
documentation. The changes include updating package manager references,
troubleshooting links, and general tool descriptions.

2.
Adjust wording in several sections to use neither Cachi2 or
Hermeto but to use generic words to keep the references to the project
as minimal as possible (in case of another rebranding).

Signed-off-by: Michal Šoltis <msoltis@redhat.com>