Mintmaker docs: point to the anchor about the rpm lockfile generation
Previously, the link in the section about "RPM lockfile with RPMs that
require subscription" was pointing to the page that completely describes
how to use entitled content in builds.

The section relevant to the lockfiles is on the bottom of the page, so
let's update the link to point to it directly.

Signed-off-by: Bruno Pimentel <bpimente@redhat.com>