fix(315): Restore numbering for custom resource partials (#318)
When moving the custom resources to partials, we had the format
following the content for the creating page which was in a list.
Maintining the indentation continuity broke the rendering when there was
no list, however (i.e. configuration as code) pages. Instead of having a
`+` to restore the indentation, we can just delete the newline. This
will ensure proper context in both types of partial inclusions.

resolves: #315

Signed-off-by: arewm <arewm@users.noreply.github.com>