fix: Document Snapshot introductory material (#324)
* fix: Document Snapshot introductory material

We just jump into working with Snapshots but we don't provide enough
context about Snapshots and what they mean. We can improve that with a
summary index.adoc.

Resolves: STONEINTG-1204

Signed-off-by: arewm <arewm@users.noreply.github.com>

* Move content from "what is a snapshot"

The `working-with-snapshots.adoc` page was previously the first page
within the Snapshots tree. It had some introductory content for
Snapshots. Since this index.adoc is now the introductory content, we
should move that here.

Signed-off-by: arewm <arewm@users.noreply.github.com>

---------

Signed-off-by: arewm <arewm@users.noreply.github.com>