Add additional patterns for managing pipeline definition
The tag and label parity pattern is simpler if there is only one
occurrence of a build pipeline to update. We already have one pattern
that can accomplish this centralization but there is another simpler
solution that may be available -- creating a separate pipeline
definition in the Component's git repository. This commit adds that as a
new pattern while also adding an additional pattern page to connect the
extracted pipeline in repo to the remote pipeline pattern.

Signed-off-by: arewm <arewm@users.noreply.github.com>