Update the getting started doc
In https://github.com/konflux-ci/docs/issues/273 we noted that the
getting started doc was kind of bumpy.

I tried to smooth it out here.

Notably, I added an explanation of Custom Resources early on that I
think was missing and I removed a good bit of text that I thought was
unnecessary and cumbersome for the introduction doc.