Add hidden machine readable content
We may want to process documentation to load it into AI models' context
windows for improving user troubleshooting. This content might not be
terribly useful for humans reading the pages.

This change creates a framework for adding machine-readable content.
While the helfulness of this content isn't currenly known, the process
show here can be repeated in the future.

Signed-off-by: arewm <arewm@users.noreply.github.com>