Update the getting started doc (#341)
* Update the getting started doc

In https://github.com/konflux-ci/docs/issues/273 we noted that the
getting started doc was kind of bumpy.

I tried to smooth it out here.

Notably, I added an explanation of Custom Resources early on that I
think was missing and I removed a good bit of text that I thought was
unnecessary and cumbersome for the introduction doc.

* Drop paragraph to reduce information overflow in getting started guide

* Link off to subsequent sections

* Drop unnecessary parenthetical

* Make Application introduce simpler and more generic

* Drop unnecessary link to single component mode from getting started docs

* Use consistent language.

In https://konflux-ci.dev/docs/releasing/, we don't use these terms but "managed environment team."

* Nest CR sections underneath the Custom Resource header

* Spell out OCI artifacts