Install RPM keys with prefetch + install root (#429)
This change adds instructions, and a reminder, on installing the
required RPM GPG keys. The YUM repo generated by hermeto enabled GPG
check (good) but it doesn't specify the location for the GPP key. It
expects any required key to already be trusted by the system.

This works in most cases. However, when using a custom install root, the
GPG keys have to be trusted in the context of the custom install root.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>