Update Key Features of Konflux (#442)
* Update Key Features of Konflux

This pull requests adds significant information to the Konflux index page.

* Make minor corrections.

This commit corrects the documentation based on team feedback.

* Remove reference to eBooks.

Remove the reference to eBooks in the Collaboration section of the document. Also, add links to the documentation that readers can click to go to the Overview and Community pages on GitHub.