doc: document behavior for override event-type (#451)
If the `pac.test.appstudio.openshift.io/event-type` annotation is not
set on a snapshot the integration service treats it as a push snapshot.
This is a point of confusion for users creating override snapshots
because in that case the event-type annotation must be set to
"override." This commit clarifies the behavior for Snapshots without the
event-type annotation.

Signed-off-by: Ryan Cole <rcyoalne@gmail.com>